refactor: install module
